+++ cotede/qctests/__init__.py
@@ -9,3 +9,4 @@ from profile_envelope import profile_envelope
 from woa_normbias import woa_normbias
 from anomaly_detection import anomaly_detection
 from possible_speed import possible_speed
+from fuzzylogic import fuzzylogic