+++ HISTORY.rst
@@ -1,12 +1,13 @@
 .. :changelog:
 
 History
--------
+=======
 
 0.14
 ----
 
 * Interface for human calibration of anomaly detection
+* Implemented fuzzy logic criteria
 
 0.13 - July, 2015
 ----